Send duplicate packets even earlier in attempt to make game smoother.
This will re-send packets as early as the next clock tick depending on conditions.  This is an experimental fix to see if this will smooth the game out even more for lossy clients because the duplicate packet has a chance of getting there in the correct frame (as opposed to the next frame).  This might ultimately mean that we are sending 2x the packets but it does seem that today's hardware can handle that just fine.